+++ packages/nx/lib/io/image_io.ml
@@ -18,6 +18,9 @@ external png_encode : Unix.file_descr -> bytes -> int -> int -> int -> unit
 external jpeg_encode : Unix.file_descr -> bytes -> int -> int -> int -> unit
   = "caml_nx_io_jpeg_encode"
 
+external png_encode_string : bytes -> int -> int -> int -> string
+  = "caml_nx_io_png_encode_string"
+
 let map_file fd size =
   if size = 0 then Array1.create int8_unsigned c_layout 0
   else
@@ -111,6 +114,10 @@ let save_png ~overwrite path data ~width ~height ~channels =
     | exception exn ->
         remove_if_exists temp;
         raise exn
+
+let encode_png data ~width ~height ~channels =
+  ignore (checked_pixels width height channels);
+  png_encode_string data width height channels
 
 let save_jpeg ~overwrite path data ~width ~height ~channels =
   ignore (checked_pixels width height channels);